fix(#sersi.dev-5): fix dark mode bug
