+++ components/Features.vue
@@ -9,18 +9,7 @@
           Define everything from file organization to dependencies, making project setup 
           consistent and automated across your team.
         </p>
-        <div class="flex justify-center items-center mb-4">
-      <div
-        class="border border-gray-200 p-5 px-8 w-auto flex justify-center items-center gap-10 rounded-sm shadow-lg"
-      >
-        <code>npx sersi<span class="text-purplish">@alpha</span> build -f sersi.yaml</code>
-        <UButton
-          icon="i-lucide-copy"
-          @click="handleCopy"
-          :color="copied ? 'neutral' : 'primary'"
-        />
-      </div>
-    </div>
+        <DependencyLabel command="build -f sersi.yaml" />
       </div>
       <div class="h-[26vh]"></div>
       <div class="h-[60vh] flex flex-col justify-center items-start gap-8">
@@ -67,18 +56,7 @@
         ease: 'easeInOut'
       }"
       >
-        <div
-          class="flex justify-between items-center h-10 w-full border-b border-gray-200 px-2"
-        >
-          <div class="flex items-center text-sm font-bold">
-            <UIcon class="size-5" name="i-simple-icons-yaml" />
-            <code>sersi.yaml</code>
-          </div>
-          <UButton size="sm" icon="i-lucide-copy" variant="outline" />
-        </div>
-        <div class="h-full w-full p-4">
-          <pre class="text-sm">{{ yamlSnippet }}</pre>
-        </div>
+       <CodeSnippet />
       </motion.div>
       <motion.div v-if="scrollValue >= 0.40 && scrollValue < 0.55"
       key="tree"
@@ -90,7 +68,7 @@
         ease: 'easeInOut'
       }"
       class="h-full w-full p-4 overflow-y-auto">
-        <UTree :items="treeData" size="lg" />
+        <FolderTree />
       </motion.div>
       <motion.div v-if="scrollValue >= 0.55">
         <CodeTypeWriter />
@@ -138,181 +116,5 @@ scaffold:
     CICD: "github-actions"
 `;
 
-const treeData = ref([
-  {
-    label: 'my-fullstack-app/',
-    defaultExpanded: true,
-    children: [
-      {
-        label: 'frontend/',
-        defaultExpanded: true,
-        children: [
-          {
-            label: '.github/',
-            children: [
-              {
-                label: 'workflows/',
-                children: [
-                  {
-                    label: 'ci.yml',
-                    icon: 'i-vscode-icons-file-type-yaml',
-                  },
-                  {
-                    label: 'release.yml',
-                    icon: 'i-vscode-icons-file-type-yaml',
-                  },
-                ],
-              },
-            ],
-          },
-          {
-            label: 'public/',
-            children: [
-              { label: 'vite.svg', icon: 'i-vscode-icons-file-type-svg' },
-            ],
-          },
-          {
-            label: 'src/',
-            children: [
-              {
-                label: 'assets/',
-                children: [
-                  { label: 'logo.svg', icon: 'i-vscode-icons-file-type-svg' },
-                ],
-              },
-              {
-                label: 'components/',
-                children: [
-                  { label: 'Hero.vue', icon: 'i-vscode-icons-file-type-vue' },
-                  { label: 'NavBar.vue', icon: 'i-vscode-icons-file-type-vue' },
-                  { label: 'Footer.vue', icon: 'i-vscode-icons-file-type-vue' },
-                ],
-              },
 
-              { label: 'main.ts', icon: 'i-vscode-icons-file-type-typescript' },
-              { label: 'styles.css', icon: 'i-vscode-icons-file-type-css' },
-              { label: 'App.vue', icon: 'i-vscode-icons-file-type-vue' },
-              { label: 'index.html', icon: 'i-vscode-icons-file-type-html' },
-            ],
-          },
-          { label: 'Dockerfile', icon: 'i-vscode-icons-file-type-docker' },
-          { label: 'package.json', icon: 'i-vscode-icons-file-type-json' },
-          { label: 'README.md', icon: 'i-vscode-icons-file-type-markdown' },
-          { label: 'sersi.yaml', icon: 'i-vscode-icons-file-type-yaml' },
-          {
-            label: 'vite.config.ts',
-            icon: 'i-vscode-icons-file-type-typescript',
-          },
-          { label: 'tsconfig.json', icon: 'i-vscode-icons-file-type-json' },
-          {
-            label: 'tsconfig.node.json',
-            icon: 'i-vscode-icons-file-type-json',
-          },
-          { label: 'tsconfig.app.json', icon: 'i-vscode-icons-file-type-json' },
-          { label: 'index.html', icon: 'i-vscode-icons-file-type-html' },
-        ],
-      },
-      {
-        label: 'backend/',
-        children: [
-          {
-            label: '.github/',
-            children: [
-              {
-                label: 'workflows/',
-                children: [
-                  {
-                    label: 'ci.yml',
-                    icon: 'i-vscode-icons-file-type-yaml',
-                  },
-                  {
-                    label: 'release.yml',
-                    icon: 'i-vscode-icons-file-type-yaml',
-                  },
-                ],
-              },
-            ],
-          },
-          {
-            label: 'src/',
-            children: [
-              { label: 'app.ts', icon: 'i-vscode-icons-file-type-typescript' },
-              {
-                label: 'server.ts',
-                icon: 'i-vscode-icons-file-type-typescript',
-              },
-              {
-                label: 'config/',
-                children: [
-                  {
-                    label: 'mongodb.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-              {
-                label: 'controllers/',
-                children: [
-                  {
-                    label: 'user.controller.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-              {
-                label: 'routes/',
-                children: [
-                  {
-                    label: 'user.route.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-              {
-                label: 'middlewares/',
-                children: [
-                  {
-                    label: 'error.middleware.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-              {
-                label: 'models/',
-                children: [
-                  {
-                    label: 'user.model.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-              {
-                label: 'services/',
-                children: [
-                  {
-                    label: 'user.service.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-              {
-                label: 'utils/',
-                children: [
-                  {
-                    label: 'logger.ts',
-                    icon: 'i-vscode-icons-file-type-typescript',
-                  },
-                ],
-              },
-            ],
-          },
-          { label: 'Dockerfile', icon: 'i-vscode-icons-file-type-docker' },
-          { label: 'package.json', icon: 'i-vscode-icons-file-type-json' },
-          { label: 'README.md', icon: 'i-vscode-icons-file-type-markdown' },
-          { label: 'tsconfig.json', icon: 'i-vscode-icons-file-type-json' },
-        ],
-      },
-    ],
-  },
-]);
 </script>
