+++ layouts/default.vue
@@ -1,8 +1,11 @@
 <template>
   <div class="px-32 md:px-64">
+    <ColorScheme placeholder="" tag="div" >
     <NavBar />
     <slot />
-  </div>
+    </ColorScheme>
+    </div>
 </template>
 
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+</script>
